# Conflicts:
#	src/main/java/com/treasure/hunt/analysis/Statistic.java
#	src/main/java/com/treasure/hunt/strategy/hider/impl/HintEvaluator.java

+++ src/main/java/com/treasure/hunt/analysis/Statistic.java
@@ -79,7 +79,7 @@ public class Statistic {
     /**
      * @return The number of {@link com.treasure.hunt.strategy.hint.Hint}'s, the {@link com.treasure.hunt.strategy.searcher.Searcher} got yet.
      */
-    public double getHintRequests() {
+    public int getHintRequests() {
         if (turns.size() > 1 && turns.get(turns.size() - 1).getHint() == null) {
             return turns.size() - 2;
         }
